+++ docs/changelog.md
@@ -14,6 +14,7 @@
 ### Features
 - The CLI now displays the list of container images currently being downloaded and validated during the Starlark
 validation step
+- `exec` now returns the command output and code
 
 ### Changes
 - Remove completion files